apps/evalgen_app/pyproject.toml
[project]
name = "evalgen_app"
version = "0.1.0"
description = "FastHTML EvalGen App"
authors = []
dependencies = [
    "evalgen",
    "databricks-sdk>=0.48.0",
    "databricks-sqlalchemy>=2.0.6",
    "fastcore>=1.8.0",
    "fastsql>=2.0.3",
    "monsterui>=1.0.21",
    "polars>=1.12.0",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "python-fasthtml>=0.12.19",
    "databricks-connect>=16.1.3",
]
requires-python = "~= 3.11"
dynamic = ["una"]  # needed for hatch-una metadata hook to work

[build-system]
requires = ["hatchling", "hatch-una"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv]
dev-dependencies = ["basedpyright>=1.29.0", "ruff>=0.11.6"]

[tool.uv.sources]


[tool.uv.sources.evalgen]
workspace = true
[tool.hatch.build.hooks.una-build]
[tool.hatch.metadata.hooks.una-meta]